Improve SNAPSHOT handling by using the artifact's base version
If a dependency has a SNAPSHOT version and comes from a remote repository, its version can be a timestamp instead of the -SNAPSHOT version. Here, inconsistencies and duplicates may occur depending on the project, which is why we now always refer to the base version.

For regular artifacts, the base version is not different from the version.