Rewrite artifact resolution to support recent Maven versions
The FSM plugin determines a Maven project for all artifacts to resolve the dependencies. The project can be located in the Maven reactor or in the local or even remote repository. Additionally, a local project can now be resolved if the plugin is running in a submodule.

For resolving artifacts from repositories, maven-artifact-transfer is now used.

For current Maven versions (3.8.8 or 3.9.x) especially the ArtifactFilter seems to be important.