Introduce 'prepare' goal as a Maven 3.9-compatible 'dependencyToXML' replacement
Newer Maven versions often caused problems with the dependencyToXML goal: the collected dependencies were either not complete or did not match the versions from the assembly plugin.

This commit introduces the new 'prepare' goal that uses the new artifact resolution and also copies dependencies to the FSM lib directory.

The old goal should not be used anymore, but they are fully backward compatible to version 1.8 to simplify migration.